# [PERF] evaluation: spreading relations with zones