--- a/src/plugins/ui_core_views/cell_evaluation/evaluator.ts
+++ b/src/plugins/ui_core_views/cell_evaluation/evaluator.ts
@@ -1,6 +1,13 @@
 import { compile } from "../../../formulas";
 import { handleError, implementationErrorMessage } from "../../../functions";
-import { lazy, positionToZone, toXC, union, unionPositionsToZone } from "../../../helpers";
+import {
+  aggregatePositionsToZones,
+  excludeTopLeft,
+  lazy,
+  positionToZone,
+  toXC,
+  union,
+} from "../../../helpers";
 import { createEvaluatedCell, evaluateLiteral } from "../../../helpers/cells";
 import { ModelConfig } from "../../../model";
 import { onIterationEndEvaluationRegistry } from "../../../registries/evaluation_registry";
@@ -56,7 +63,8 @@ export class Evaluator {
   }
 
   getSpreadZone(position: CellPosition, options = { ignoreSpillError: false }): Zone | undefined {
-    if (!this.spreadingRelations.isArrayFormula(position)) {
+    const spreadZone = this.spreadingRelations.getArrayResultZone(position);
+    if (!spreadZone) {
       return undefined;
     }
     const evaluatedCell = this.evaluatedCells.get(position);
@@ -66,8 +74,7 @@ export class Evaluator {
     ) {
       return positionToZone(position);
     }
-    const spreadPositions = Array.from(this.spreadingRelations.getArrayResultPositions(position));
-    return union(positionToZone(position), unionPositionsToZone(spreadPositions));
+    return union(positionToZone(position), spreadZone);
   }
 
   getEvaluatedPositions(): CellPosition[] {
@@ -79,10 +86,16 @@ export class Evaluator {
   }
 
   getArrayFormulaSpreadingOn(position: CellPosition): CellPosition | undefined {
-    if (!this.spreadingRelations.hasArrayFormulaResult(position)) {
+    const hasArrayFormulaResult =
+      this.getEvaluatedCell(position).type !== CellValueType.empty &&
+      !this.getters.getCell(position)?.isFormula;
+    if (!hasArrayFormulaResult) {
       return this.spreadingRelations.isArrayFormula(position) ? position : undefined;
     }
-    const arrayFormulas = this.spreadingRelations.getFormulaPositionsSpreadingOn(position);
+    const arrayFormulas = this.spreadingRelations.getFormulaPositionsSpreadingOn(
+      position.sheetId,
+      positionToZone(position)
+    );
     return Array.from(arrayFormulas).find((position) => !this.blockedArrayFormulas.has(position));
   }
 
@@ -159,7 +172,12 @@ export class Evaluator {
         impactedPositions.add(position);
       }
     }
-    impactedPositions.addMany(this.getArrayFormulasBlockedBy(impactedPositions));
+    const zonesBySheetIds = aggregatePositionsToZones(impactedPositions);
+    for (const sheetId in zonesBySheetIds) {
+      for (const zone of zonesBySheetIds[sheetId]) {
+        impactedPositions.addMany(this.getArrayFormulasBlockedBy(sheetId, zone));
+      }
+    }
     return impactedPositions;
   }
 
@@ -228,19 +246,16 @@ export class Evaluator {
    * Return the position of formulas blocked by the given positions
    * as well as all their dependencies.
    */
-  private getArrayFormulasBlockedBy(positions: Iterable<CellPosition>): Iterable<CellPosition> {
+  private getArrayFormulasBlockedBy(sheetId: UID, zone: Zone): Iterable<CellPosition> {
     const arrayFormulaPositions = this.createEmptyPositionSet();
-    for (const position of positions) {
-      if (!this.spreadingRelations.hasArrayFormulaResult(position)) {
-        continue;
-      }
-      const arrayFormulas = this.spreadingRelations.getFormulaPositionsSpreadingOn(position);
-      arrayFormulaPositions.addMany(arrayFormulas);
-      const arrayFormulaPosition = this.getArrayFormulaSpreadingOn(position);
-      if (arrayFormulaPosition) {
-        // ignore the formula spreading on the position. Keep only the blocked ones
-        arrayFormulaPositions.delete(arrayFormulaPosition);
-      }
+    const arrayFormulas = this.spreadingRelations.getFormulaPositionsSpreadingOn(sheetId, zone);
+    arrayFormulaPositions.addMany(arrayFormulas);
+    const spilledPositions = [...arrayFormulas].filter(
+      (position) => !this.blockedArrayFormulas.has(position)
+    );
+    if (spilledPositions.length) {
+      // ignore the formula spreading on the position. Keep only the blocked ones
+      arrayFormulaPositions.deleteMany(spilledPositions);
     }
     arrayFormulaPositions.addMany(this.getCellsDependingOn(arrayFormulaPositions));
     return arrayFormulaPositions;
@@ -340,7 +355,13 @@ export class Evaluator {
     const nbColumns = formulaReturn.length;
     const nbRows = formulaReturn[0].length;
 
-    forEachSpreadPositionInMatrix(nbColumns, nbRows, this.updateSpreadRelation(formulaPosition));
+    const resultZone = {
+      top: formulaPosition.row,
+      bottom: formulaPosition.row + nbRows - 1,
+      left: formulaPosition.col,
+      right: formulaPosition.col + nbColumns - 1,
+    };
+    this.spreadingRelations.addRelation({ resultZone, arrayFormulaPosition: formulaPosition });
     this.assertNoMergedCellsInSpreadZone(formulaPosition, formulaReturn);
     forEachSpreadPositionInMatrix(nbColumns, nbRows, this.checkCollision(formulaPosition));
     forEachSpreadPositionInMatrix(
@@ -349,7 +370,7 @@ export class Evaluator {
       // thanks to the isMatrix check above, we know that formulaReturn is MatrixFunctionReturn
       this.spreadValues(formulaPosition, formulaReturn)
     );
-    this.invalidatePositionsDependingOnSpread(formulaPosition, nbColumns, nbRows);
+    this.invalidatePositionsDependingOnSpread(formulaPosition.sheetId, resultZone);
     return createEvaluatedCell(
       nullValueToZeroValue(formulaReturn[0][0]),
       this.getters.getLocale(),
@@ -357,32 +378,11 @@ export class Evaluator {
     );
   }
 
-  private invalidatePositionsDependingOnSpread(
-    arrayFormulaPosition: CellPosition,
-    nbColumns: number,
-    nbRows: number
-  ) {
+  private invalidatePositionsDependingOnSpread(sheetId: UID, resultZone: Zone) {
     // the result matrix is split in 2 zones to exclude the array formula position
-    const top = arrayFormulaPosition.row;
-    const left = arrayFormulaPosition.col;
-    const bottom = top + nbRows - 1;
-    const leftColumnZone = {
-      top: top + 1,
-      bottom,
-      left,
-      right: left,
-    };
-    const rightPartZone = {
-      top,
-      bottom,
-      left: left + 1,
-      right: left + nbColumns - 1,
-    };
-    const sheetId = arrayFormulaPosition.sheetId;
-    const invalidatedPositions = this.formulaDependencies().getCellsDependingOn([
-      { sheetId, zone: rightPartZone },
-      { sheetId, zone: leftColumnZone },
-    ]);
+    const invalidatedPositions = this.formulaDependencies().getCellsDependingOn(
+      excludeTopLeft(resultZone).map((zone) => ({ sheetId, zone }))
+    );
     this.nextPositionsToUpdate.addMany(invalidatedPositions);
   }
 
@@ -436,19 +436,6 @@ export class Evaluator {
     );
   }
 
-  private updateSpreadRelation({
-    sheetId,
-    col,
-    row,
-  }: CellPosition): (i: number, j: number) => void {
-    const arrayFormulaPosition = { sheetId, col, row };
-    const updateSpreadRelation = (i: number, j: number) => {
-      const position = { sheetId, col: i + col, row: j + row };
-      this.spreadingRelations.addRelation({ resultPosition: position, arrayFormulaPosition });
-    };
-    return updateSpreadRelation;
-  }
-
   private checkCollision(formulaPosition: CellPosition): (i: number, j: number) => void {
     const { sheetId, col, row } = formulaPosition;
     const checkCollision = (i: number, j: number) => {
@@ -489,22 +476,25 @@ export class Evaluator {
   }
 
   private invalidateSpreading(position: CellPosition) {
-    if (!this.spreadingRelations.isArrayFormula(position)) {
+    const zone = this.spreadingRelations.getArrayResultZone(position);
+    if (!zone) {
       return;
     }
-    const invalidated = this.createEmptyPositionSet();
-    for (const child of this.spreadingRelations.getArrayResultPositions(position)) {
-      const content = this.getters.getCell(child)?.content;
-      if (content) {
-        // there's no point at re-evaluating overlapping array formulas,
-        // there's still a collision
-        continue;
+    for (let col = zone.left; col <= zone.right; col++) {
+      for (let row = zone.top; row <= zone.bottom; row++) {
+        const resultPosition = { sheetId: position.sheetId, col, row };
+        const content = this.getters.getCell(resultPosition)?.content;
+        if (content) {
+          // there's no point at re-evaluating overlapping array formulas,
+          // there's still a collision
+          continue;
+        }
+        this.evaluatedCells.delete(resultPosition);
       }
-      invalidated.add(child);
-      this.evaluatedCells.delete(child);
     }
-    this.nextPositionsToUpdate.addMany(this.getCellsDependingOn(invalidated));
-    this.nextPositionsToUpdate.addMany(this.getArrayFormulasBlockedBy(invalidated));
+    const sheetId = position.sheetId;
+    this.invalidatePositionsDependingOnSpread(sheetId, zone);
+    this.nextPositionsToUpdate.addMany(this.getArrayFormulasBlockedBy(sheetId, zone));
   }
 
   // ----------------------------------------------------------