@robodoo 2024-07-20
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/4692.png)](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/4692)